function-predict/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.22.3
typing-extensions==4.1.1

# Torch for running the function locally
torch==1.10.0

# Torch for deploying the function on Azure
# -f https://download.pytorch.org/whl/torch_stable.html
# torch==1.10.0+cpu